Change format for wedge angle
Originally I had it so that it would return a 0 where there were no wedge angles. This was because for case 2 in the truth table it said '30,0,30' but the second beam has no wedge however i think its better to make it clear and say no wedge cause otherwise it would be represented the same as a wedge angle of 0 degrees. Now its changed to saying 'no wedge' in these situations