Give a PASS/FAIL when VMAT file
The VMAT files only have one beam but in this one beam each control point has a different gantry and ssd. To deal with this when VMAT extract gantry retrieves a (extremely long) list of each gantry for each control point. The SSD does the same. The truth table values give a list of gantrys and an equally long list of SSD values. To verify whether the extracted values match the truth table values the SSD values and Gantry angles need to co occur the same way they do in the truth table. For instance if the 3rd gantry value is 180 and the 3rd SSD is 75 that would mean that for the VMAT a control point that has a gantry of 180 would need to have an SSD of (roughly) 75. Im pretty sure this is the correct interpretation but I will verify.

Hopefully that makes sense.